💄 refactor: 移除Navbar.tsx，將其整合至新的Navbar組件，並更新相關引用
