+++ src/components/section/Home/SkillGrid.tsx
@@ -99,7 +99,7 @@ const skillIcons: Skill[] = [
   {
     name: "Nginx",
     fileName: "nginx_symbol.svg",
-    gradientColors: ["rgba(97, 218, 251, 0.4)"],
+    gradientColors: ["rgba(0, 153, 0, 0.4)"],
     position: { row: 1, col: 5 },
   },
 ];
