+++ src/pages/Home.tsx
@@ -1,5 +1,5 @@
 import { HeroSection } from "../components/section/Home/HeroSection";
-import { Navbar } from "../components/section/Home/Navbar";
+import { Navbar } from "../components/Navbar";
 import { SkillGrid } from "../components/section/Home/SkillGrid";
 
 export const Home = () => {
